requirements.txt
ace_tools==0.0
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
attrs==25.3.0
ccxt==4.4.86
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
fonttools==4.58.1
frozenlist==1.6.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
mplfinance==0.12.10b0
multidict==6.4.4
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
ta==0.11.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
